test(normalize): move inline test modules to sibling _tests.rs files
Extract the #[cfg(test)] modules from the five Composio normalizer files
(clickup, github, helpers, linear, notion) into per-file <name>_tests.rs
siblings, matching the repo convention of keeping tests out of
implementation files.
